refactor(): add command on new structure (unfinished)
