+++ lib/actions/index.ts
@@ -1,2 +1,6 @@
 export * from './abstract.action';
+export * from './build.action';
+export * from './update.action';
+export * from './update-project.action';
+export * from './start.action';
 export * from './version.action';
