+++ lib/actions/utils/update-project/update-babelrc.ts
