+++ lib/actions/utils/update-project/update-plugins-config.ts
