+++ lib/bin/tars.ts
@@ -4,7 +4,6 @@ import { CommanderStatic } from 'commander';
 import { CommandLoader } from '../commands';
 import * as fs from 'fs';
 import * as path from 'path';
-import { isTarsInited, tarsNotInitedActions } from '../utils';
 
 const args = process.argv.slice(2);
 const cliRootPath = path.resolve(__dirname, '../../');
@@ -20,148 +19,11 @@ try {
 process.env.cliRoot = cliRootPath;
 process.env.npmRoot = npmRootPath;
 
-/**
- * Check TARS initialization and tars-config.js in current directory
- * @return {Boolean} TARS init status
- */
-const isTarsReadyToWork = (): boolean => {
-  const { inited, error } = isTarsInited();
-
-  // If we are not in TARS directory or TARS has not been inited
-  if (!inited) {
-    if (!error) {
-      tarsNotInitedActions();
-    }
-    return false;
-  }
-  return true;
-}
-
 const program: CommanderStatic = commander;
 
 program.usage(
   '[command] [options] \n         Command without flags will be started in interactive mode.',
 );
-
-program
-  .command('init')
-  .description('Init TARS-project in current directory')
-  .option('--exclude-html', 'Prevent templater-files uploading')
-  .option('--exclude-css', 'Prevent preprocessor-files uploading')
-  .option('--silent', 'TARS will not ask any question about configuration')
-  .option('-s, --source <source>', 'Change source of TARS')
-  .action(options => require('../command-actions/init')(options));
-
-program
-  .command('build')
-  .alias('bld')
-  .description('Build project without watchers')
-  .option('-r, --release', 'Create release build')
-  .option('-m, --min', 'Create build with minified files')
-  .option('--ie9', 'Generate files for ie9')
-  .option('--ie8', 'Generate files for ie8')
-  .option('--ie', 'Generate files for ie')
-  .option('--silent', 'Start build in silent mode, without promt')
-  .option('--custom-flags <customFlags>', 'Add custom flags')
-  .action(options => {
-    if (isTarsReadyToWork()) {
-      require('../command-actions/build')(options);
-    }
-  });
-
-program
-  .command('dev')
-  .alias('development')
-  .description('Build project with watchers')
-  .option('-t, --tunnel', 'Create tunnel to the Internet')
-  .option('-l, --livereload', 'Start server')
-  .option('--lr', 'Allias for livereload')
-  .option('--ie9', 'Generate files for ie9')
-  .option('--ie8', 'Generate files for ie8')
-  .option('--ie', 'Generate files for ie')
-  .option('--silent', 'Start dev in silent mode, without promt')
-  .option('--custom-flags <customFlags>', 'Add custom flags')
-  .action(options => {
-    if (isTarsReadyToWork()) {
-      require('../command-actions/dev')(options);
-    }
-  });
-
-program
-  .command('add-component <componentName>')
-  .alias('add-module')
-  .description('Add component to markup/components directory')
-  .option(
-    '-b, --basic',
-    'Add component with .js, .scss (.less, .styl) and .html (.jade) files',
-  )
-  .option(
-    '-a, --assets',
-    'Add component with general files + folder for assets',
-  )
-  .option('-d, --data', 'Add component with general files + folder for data')
-  .option('-i, --ie', 'Add component with general files + folder for IE')
-  .option(
-    '-f, --full',
-    'Add component with all files and folders (assets folder, folder for IE and so on)',
-  )
-  .option('-t, --template', 'Add component as a copy of _template component')
-  .option(
-    '-s, --scheme [schemeFile]',
-    'Add component, which structure is based on scheme file',
-  )
-  .option('-e, --empty', 'Add component without files')
-  .option('--custom-path <customPath>', 'Add component into custom folder')
-  .option('--silent', 'Add component in silent mode, without promt')
-  .action((componentName, options) => {
-    if (isTarsReadyToWork()) {
-      require('../command-actions/add-component')(componentName, options);
-    }
-  });
-
-program
-  .command('add-page <pageName>')
-  .description('Add page to markup/pages directory')
-  .option('-e, --empty', 'Add empty file')
-  .action((pageName, options) => {
-    if (isTarsReadyToWork()) {
-      require('../command-actions/add-page')(pageName, options);
-    }
-  });
-
-program
-  .command('update')
-  .alias('upgrade')
-  .description('Update TARS-cli')
-  .action(() => require('../command-actions/update')());
-
-program
-  .command('update-project')
-  .alias('upgrade-project')
-  .description('Update TARS in current project')
-  .option('-f, --force', 'Force update, even you have the latest version')
-  .option('--exclude-html', 'Prevent templater-files updating')
-  .option('--exclude-css', 'Prevent preprocessor-files updating')
-  .option('-s, --source <source>', 'Change source of TARS for updating')
-  .action(options => {
-    if (isTarsReadyToWork()) {
-      require('../command-actions/update-project')(options);
-    }
-  });
-
-program
-  .command('start <taskName>')
-  .alias('run')
-  .description('Start task from the local gulpfile')
-  .option(
-    '--flags <flags>',
-    'Add flags "--flags" \'flags, with space separator\'',
-  )
-  .action((taskName, options) => {
-    if (isTarsReadyToWork()) {
-      require('../command-actions/start-task')(taskName, options);
-    }
-  });
 
 if (!args.length) {
   program.outputHelp();
