+++ lib/commands/command.loader.ts
@@ -1,9 +1,15 @@
 import { CommanderStatic } from 'commander';
-import { VersionAction } from '../actions';
+import { UpdateAction, UpdateProjectAction, StartAction, VersionAction } from '../actions';
+import { UpdateCommand } from './update.command';
+import { UpdateProjectCommand } from './update-project.command';
+import { StartCommand } from './start.command';
 import { VersionCommand } from './version.command';
 
 export class CommandLoader {
   public static load(program: CommanderStatic): void {
+    new UpdateCommand(new UpdateAction()).load(program);
+    new UpdateProjectCommand(new UpdateProjectAction()).load(program);
+    new StartCommand(new StartAction()).load(program);
     new VersionCommand(new VersionAction()).load(program);
 
     this.handleInvalidCommand(program);
